Modificaciones ABMC Familia - Correccion de duplicado al editar/eliminar integrantes

+++ src/main/java/desi/tp/accesoDatos/AsistidoRepo.java
@@ -9,5 +9,5 @@ import desi.tp.entidades.Asistido;
 
 public interface AsistidoRepo extends JpaRepository <Asistido, Integer> {
 
-	   List<Asistido> findByDniIn(List<Integer> listaDni);
+	   List<Asistido> findByDniIn(List<Long> listaDni);
 }
